+++ jabgui/src/main/java/org/jabref/gui/actions/StandardActions.java
@@ -87,7 +87,7 @@ public enum StandardActions implements Action {
     REPLACE_ALL(Localization.lang("Find and replace"), KeyBinding.REPLACE_STRING),
     MANAGE_KEYWORDS(Localization.lang("Manage keywords")),
     MASS_SET_FIELDS(Localization.lang("Manage field names & content")),
-    MERGE_BIB_FILES_INTO_CURRENT_BIB(Localization.lang("Merge other bib files into current library...")),
+    MERGE_BIB_FILES_INTO_CURRENT_BIB(Localization.lang("Merge other BibTeX files into current library...")),
 
     AUTOMATIC_FIELD_EDITOR(Localization.lang("Automatic field editor")),
     TOGGLE_GROUPS(Localization.lang("Groups"), IconTheme.JabRefIcons.TOGGLE_GROUPS, KeyBinding.TOGGLE_GROUPS_INTERFACE),